refactor: rename and extract correct data for the home page and timeline list
